Fix compiler warning (real problem on 64-bit).
git-svn-id: http://svn.python.org/projects/python/branches/py3k-buffer@56948 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Include/object.h
@@ -151,7 +151,7 @@ typedef struct bufferinfo {
         Py_ssize_t *shape;
         Py_ssize_t *strides;
         Py_ssize_t *suboffsets;
-        int itemsize;
+        Py_ssize_t itemsize;
         void *internal;
 } PyBuffer;
